Revert "minor fix"
This reverts commit d3937f4ed14991d9f449f4a510616e7c7ce0597a.
