+++ pkg/metrics/series_test.go
@@ -386,3 +386,43 @@ func BenchmarkPopulateDeviceFieldWithDevice(b *testing.B) {
 		populateDeviceField(s)
 	}
 }
+
+func TestSortAndGroupSeries(t *testing.T) {
+	series := Series{
+		{
+			Points: []Point{
+				{Ts: 12345.0, Value: float64(21.21)},
+				{Ts: 67890.0, Value: float64(12.12)},
+			},
+			MType:    APIGaugeType,
+			Name:     "test.metrics",
+			Interval: 15,
+			Tags:     []string{"tag1", "tag2:yes"},
+		},
+		{
+			Points: []Point{
+				{Ts: 12345.0, Value: float64(21.21)},
+				{Ts: 67890.0, Value: float64(12.12)},
+			},
+			MType:    APIRateType,
+			Name:     "test.metrics",
+			Interval: 15,
+			Tags:     []string{"tag1", "tag2:no"},
+		},
+		{
+			Points:   []Point{},
+			MType:    APICountType,
+			Name:     "test.metrics2",
+			Interval: 15,
+			Tags:     nil,
+		},
+	}
+
+	expectedGroups := GroupedSeries{
+		{series[0], series[1]},
+		{series[2]},
+	}
+
+	groups := sortAndGroupSeries(series)
+	assert.Equal(t, expectedGroups, groups)
+}
